GenericNetlinkFamily

+++ NetlinkCore/Interop/Generic/Constants.cs
@@ -2,5 +2,5 @@ namespace NetlinkCore.Interop.Generic;
 
 internal static class Constants
 {
-    public static int GENL_ID_CTRL = 0x10;
+    public static ushort GENL_ID_CTRL = 0x10;
 }
